refactor(tests): nest capture, gpu and device service tests by domain
