+++ tests/unit/renderer/infrastructure/services/devices/trpc-device-status.port.test.ts
@@ -1,14 +1,14 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 
 vi.mock('@renderer/infrastructure/ipc/trpc-client', async () => ({
-  trpcClient: (await import('../../../../support/mocks/trpc-client.mock')).createTrpcClientMock()
+  trpcClient: (await import('../../../../../support/mocks/trpc-client.mock')).createTrpcClientMock()
 }));
 
 import { TrpcDeviceStatusPort } from '@renderer/infrastructure/services/devices/trpc-device-status.port';
 import { trpcClient } from '@renderer/infrastructure/ipc/trpc-client';
-import { emitTrpcData } from '../../../../support/mocks/trpc-client.mock';
-import { createLoggerFactory } from '../../../../factories/index.js';
-import { createChromaticDeviceStatusPayload, createChromaticDeviceInfoPayload } from '../../../../devices/media.testkit';
+import { emitTrpcData } from '../../../../../support/mocks/trpc-client.mock';
+import { createLoggerFactory } from '../../../../../factories/index.js';
+import { createChromaticDeviceStatusPayload, createChromaticDeviceInfoPayload } from '../../../../../devices/media.testkit';
 
 describe('TrpcDeviceStatusPort', () => {
   let port: TrpcDeviceStatusPort;
